requirements.txt
flask==2.2.3
flask-sqlalchemy==3.0.3
flask-migrate==4.0.4
flask-jwt-extended==4.4.4
pymysql==1.0.3
python-dotenv==1.0.0
bcrypt==4.0.1
marshmallow==3.19.0
gunicorn==20.1.0
pytest==7.3.1
flasgger==0.9.5
cryptography==39.0.1

